[********]统一json解析软件

Created-by: insightDeveloper

Match-id-f32ed15e5626015866647cd053e5c210e0cee8fc

+++ server/src/modules/memory/protocol/MemoryProtocolRequest.h
@@ -16,8 +16,8 @@ namespace Protocol {
 struct MemoryOperatorParams {
     std::string rankId;
     std::string orderName;
-    double minSize;
-    double maxSize;
+    int64_t minSize;
+    int64_t maxSize;
     double startTime;
     double endTime;
     int64_t currentPage = 0;
